+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 3.28.4 - 2025-08-11
+
+-   Fix `startAfterDocument` in YustDatabaseService
+
+## 3.28.3 - 2025-08-08
+
+-   Fix `isIso8601String` extension
+
+## 3.28.2 - 2025-08-06
+
+-   Improve Number Formatting
+-   Improve Mocked DB Service
+-   Improve Date Checking
+
 ## 3.28.1 - 2025-07-14
 
 -   Add support for Time-To-Live (via expireAt) in YustDoc